Putting the task pool in its own file.
